cleanup asset creation
